Build js

+++ django_file_form/static/file_form/file_form.debug.js
@@ -4454,9 +4454,9 @@ var upload_file_UploadFile = /*#__PURE__*/function () {
     var _this = this;
 
     var callbacks = _ref.callbacks,
-        _fieldName = _ref.fieldName,
+        fieldName = _ref.fieldName,
         form = _ref.form,
-        _formId = _ref.formId,
+        formId = _ref.formId,
         initial = _ref.initial,
         input = _ref.input,
         multiple = _ref.multiple,
@@ -4466,7 +4466,7 @@ var upload_file_UploadFile = /*#__PURE__*/function () {
         skipRequired = _ref.skipRequired,
         supportDropArea = _ref.supportDropArea,
         translations = _ref.translations,
-        _uploadUrl = _ref.uploadUrl;
+        uploadUrl = _ref.uploadUrl;
 
     classCallCheck_default()(this, UploadFile);
 
@@ -4494,6 +4494,8 @@ var upload_file_UploadFile = /*#__PURE__*/function () {
 
     defineProperty_default()(this, "uploads", void 0);
 
+    defineProperty_default()(this, "uploadStatuses", void 0);
+
     defineProperty_default()(this, "uploadFiles", function (files) {
       if (files.length === 0) {
         return;
@@ -4506,73 +4508,7 @@ var upload_file_UploadFile = /*#__PURE__*/function () {
       }
 
       files.forEach(function (file) {
-        var fieldName = _this.fieldName,
-            formId = _this.formId,
-            renderer = _this.renderer,
-            uploads = _this.uploads,
-            uploadUrl = _this.uploadUrl;
-        var filename = file.name;
-        var uploadIndex = uploads.length; // #323 remove existing file
-
-        for (var index = 0; index < _this.uploads.length; ++index) {
-          var existingUpload = _this.uploads[index];
-
-          if (existingUpload instanceof browser_Upload) {
-            var _existingUpload$optio, _existingUpload$optio2;
-
-            if (((_existingUpload$optio = existingUpload.options) === null || _existingUpload$optio === void 0 ? void 0 : (_existingUpload$optio2 = _existingUpload$optio.metadata) === null || _existingUpload$optio2 === void 0 ? void 0 : _existingUpload$optio2.filename) === filename) {
-              uploadIndex = index;
-
-              var el = _this.renderer.findFileDiv(index);
-
-              if (el.classList.contains("dff-upload-fail")) {
-                _this.deleteUpload(index);
-              } else if (el.classList.contains("dff-upload-success")) {
-                _this.deleteFromServer(index);
-              } else {
-                void existingUpload.abort(true);
-
-                _this.deleteUpload(index);
-              }
-
-              break;
-            }
-          } else if (existingUpload) {
-            if (existingUpload.name === filename) {
-              _this.deletePlaceholder(index);
-
-              uploadIndex = index;
-              break;
-            }
-          }
-        }
-
-        var upload = new browser_Upload(file, {
-          endpoint: uploadUrl,
-          metadata: {
-            fieldName: fieldName,
-            filename: filename,
-            formId: formId
-          },
-          onError: function onError(error) {
-            return _this.handleError(uploadIndex, error);
-          },
-          onProgress: function onProgress(bytesUploaded, bytesTotal) {
-            return _this.handleProgress(uploadIndex, bytesUploaded, bytesTotal);
-          },
-          onSuccess: function onSuccess() {
-            return _this.handleSuccess(uploadIndex, upload.file.size);
-          },
-          retryDelays: _this.retryDelays || [0, 1000, 3000, 5000]
-        });
-        upload.start();
-        renderer.addNewUpload(filename, uploadIndex);
-
-        if (uploadIndex == _this.uploads.length) {
-          _this.uploads.push(upload);
-        } else {
-          _this.uploads[uploadIndex] = upload;
-        }
+        _this.uploadFile(file);
       });
 
       _this.checkDropHint();
@@ -4599,7 +4535,7 @@ var upload_file_UploadFile = /*#__PURE__*/function () {
         var uploadIndex = getUploadIndex();
 
         if (uploadIndex !== null) {
-          _this.handleDelete(uploadIndex);
+          _this.removeExistingUpload(uploadIndex);
         }
 
         e.preventDefault();
@@ -4633,6 +4569,7 @@ var upload_file_UploadFile = /*#__PURE__*/function () {
     defineProperty_default()(this, "handleError", function (uploadIndex, error) {
       _this.renderer.setError(uploadIndex);
 
+      _this.uploadStatuses[uploadIndex] = "error";
       var onError = _this.callbacks.onError;
 
       if (onError) {
@@ -4648,6 +4585,7 @@ var upload_file_UploadFile = /*#__PURE__*/function () {
       var renderer = _this.renderer;
       renderer.clearInput();
       renderer.setSuccess(uploadIndex, uploadedSize);
+      _this.uploadStatuses[uploadIndex] = "done";
       var onSuccess = _this.callbacks.onSuccess;
 
       if (onSuccess) {
@@ -4660,16 +4598,17 @@ var upload_file_UploadFile = /*#__PURE__*/function () {
     });
 
     this.callbacks = callbacks;
-    this.fieldName = _fieldName;
+    this.fieldName = fieldName;
     this.form = form;
-    this.formId = _formId;
+    this.formId = formId;
     this.multiple = multiple;
     this.prefix = prefix;
     this.retryDelays = retryDelays;
     this.supportDropArea = supportDropArea;
-    this.uploadUrl = _uploadUrl;
+    this.uploadUrl = uploadUrl;
     this.uploadIndex = 0;
     this.uploads = [];
+    this.uploadStatuses = [];
     this.renderer = new render_upload_file({
       parent: parent,
       input: input,
@@ -4727,6 +4666,8 @@ var upload_file_UploadFile = /*#__PURE__*/function () {
             url: url
           });
         }
+
+        _this2.uploadStatuses.push("done");
       };
 
       if (multiple) {
@@ -4740,13 +4681,103 @@ var upload_file_UploadFile = /*#__PURE__*/function () {
       }
     }
   }, {
-    key: "handleDelete",
-    value: function handleDelete(uploadIndex) {
+    key: "uploadFile",
+    value: function uploadFile(file) {
+      var _this3 = this;
+
+      var fieldName = this.fieldName,
+          formId = this.formId,
+          renderer = this.renderer,
+          uploads = this.uploads,
+          uploadUrl = this.uploadUrl;
+      var filename = file.name;
+      var existingUploadIndex = this.findUpload(filename);
+
+      if (existingUploadIndex !== null) {
+        this.removeExistingUpload(existingUploadIndex);
+      }
+
+      var uploadIndex = uploads.length;
+      var upload = new browser_Upload(file, {
+        endpoint: uploadUrl,
+        metadata: {
+          fieldName: fieldName,
+          filename: filename,
+          formId: formId
+        },
+        onError: function onError(error) {
+          return _this3.handleError(uploadIndex, error);
+        },
+        onProgress: function onProgress(bytesUploaded, bytesTotal) {
+          return _this3.handleProgress(uploadIndex, bytesUploaded, bytesTotal);
+        },
+        onSuccess: function onSuccess() {
+          return _this3.handleSuccess(uploadIndex, upload.file.size);
+        },
+        retryDelays: this.retryDelays || [0, 1000, 3000, 5000]
+      });
+      upload.start();
+      renderer.addNewUpload(filename, uploadIndex);
+      this.uploads.push(upload);
+      this.uploadStatuses.push("uploading");
+    }
+  }, {
+    key: "findUpload",
+    value: function findUpload(filename) {
+      var upload = this.uploads.find(function (upload) {
+        if (upload instanceof browser_Upload) {
+          var _upload$options, _upload$options$metad;
+
+          return ((_upload$options = upload.options) === null || _upload$options === void 0 ? void 0 : (_upload$options$metad = _upload$options.metadata) === null || _upload$options$metad === void 0 ? void 0 : _upload$options$metad.filename) === filename;
+        } else if (upload) {
+          return upload.name === filename;
+        } else {
+          return false;
+        }
+      });
+      var index = this.uploads.indexOf(upload);
+      return index >= 0 ? index : null;
+    }
+  }, {
+    key: "removeExistingUpload",
+    value: function removeExistingUpload(uploadIndex) {
+      var uploadStatus = this.uploadStatuses[uploadIndex];
+
+      if (uploadStatus === "uploading") {
+        var _upload4 = this.uploads[uploadIndex];
+        void _upload4.abort(true);
+      }
+
       var upload = this.uploads[uploadIndex];
 
+      if (!upload) {
+        return;
+      }
+
       if (upload instanceof browser_Upload || upload.url) {
-        this.deleteFromServer(uploadIndex);
-      } else {
+        var _uploadStatus = this.uploadStatuses[uploadIndex];
+
+        switch (_uploadStatus) {
+          case "done":
+            {
+              this.deleteFromServer(uploadIndex);
+              break;
+            }
+
+          case "error":
+            {
+              this.deleteUpload(uploadIndex);
+              break;
+            }
+
+          case "uploading":
+            {
+              var _upload5 = this.uploads[uploadIndex];
+              void _upload5.abort(true);
+              break;
+            }
+        }
+      } else if (upload.placeholder) {
         this.deletePlaceholder(uploadIndex);
       }
     }
@@ -4754,6 +4785,11 @@ var upload_file_UploadFile = /*#__PURE__*/function () {
     key: "deleteUpload",
     value: function deleteUpload(uploadIndex) {
       var upload = this.uploads[uploadIndex];
+
+      if (!upload) {
+        return;
+      }
+
       this.renderer.deleteFile(uploadIndex);
       delete this.uploads[uploadIndex];
       this.checkDropHint();
@@ -4766,10 +4802,10 @@ var upload_file_UploadFile = /*#__PURE__*/function () {
   }, {
     key: "deleteFromServer",
     value: function deleteFromServer(uploadIndex) {
-      var _this3 = this;
+      var _this4 = this;
 
       var upload = this.uploads[uploadIndex];
-      var url = upload.url;
+      var url = upload === null || upload === void 0 ? void 0 : upload.url;
 
       if (!url) {
         return;
@@ -4781,9 +4817,9 @@ var upload_file_UploadFile = /*#__PURE__*/function () {
 
       xhr.onload = function () {
         if (xhr.status === 204) {
-          _this3.deleteUpload(uploadIndex);
+          _this4.deleteUpload(uploadIndex);
         } else {
-          _this3.renderer.setDeleteFailed(uploadIndex);
+          _this4.renderer.setDeleteFailed(uploadIndex);
         }
       };
 
@@ -4836,7 +4872,7 @@ var upload_file_UploadFile = /*#__PURE__*/function () {
     key: "updatePlaceholderInput",
     value: function updatePlaceholderInput() {
       var placeholdersInfo = this.uploads.filter(function (upload) {
-        return !(upload instanceof browser_Upload) && upload.placeholder;
+        return !(upload instanceof browser_Upload) && (upload === null || upload === void 0 ? void 0 : upload.placeholder);
       });
       var input = Object(util["a" /* findInput */])(this.form, Object(util["e" /* getPlaceholderFieldName */])(this.fieldName, this.prefix), this.prefix);
 
@@ -5122,7 +5158,7 @@ function encode(input) {
  * @api public
  */
 function querystring(query) {
-  var parser = /([^=?#&]+)=?([^&]*)/g
+  var parser = /([^=?&]+)=?([^&]*)/g
     , result = {}
     , part;
 
@@ -5177,8 +5213,8 @@ function querystringify(obj, prefix) {
         value = '';
       }
 
-      key = encode(key);
-      value = encode(value);
+      key = encodeURIComponent(key);
+      value = encodeURIComponent(value);
 
       //
       // If we failed to encode the strings, we should bail out as we don't
